requirements.txt
selenium
webdriver-manager
pandas
reportlab
rpa
python-dotenv